risc-v

+++ cctest.sh
@@ -1,9 +1,18 @@
-DIR="$1/"
+TARGET="$1"
+DIR="$2/"
+
+if [ "$TARGET" = "" ]; then
+	TARGET="x8664"
+fi
+if [ "$DIR" = "/" ]; then
+	echo "Usage: sh test.sh [x8664/riscv] [dir]"
+	exit 1
+fi
 
 cd 9cc
 echo "`pwd`"
 
-NCC="../9cc/9cc"
+NCC="../9cc/9cc --arch $TARGET "
 PRPR="../prpr/prpr --stddir ../std/"
 
 rm -rf $DIR
@@ -37,7 +46,6 @@ mgcc() {
 
 compile "charutil"
 compile "fileutil"
-compile "gen"
 compile "list"
 compile "main"
 compile "parse"
@@ -49,11 +57,13 @@ compile "varutil"
 compile "analyze"
 compile "mymath"
 compile "il"
+compile "abst"
+compile "gen_x8664"
+compile "gen_riscv"
 compile "error"
 
 #mgcc "charutil"
 #mgcc "fileutil"
-#mgcc "gen"
 #mgcc "list"
 #mgcc "main"
 #mgcc "parse"
@@ -65,7 +75,10 @@ compile "error"
 #mgcc "analyze"
 #mgcc "mymath"
 #mgcc "il"
+#mgcc "abst"
+#mgcc "gen_x8664"
+#mgcc "get_riscv"
 #mgcc "error"
 
 cd $DIR
-gcc -o ../9cc charutil.s fileutil.s gen.s list.s main.s parse.s parseutil.s stack.s tokenize.s typeutil.s varutil.s analyze.s mymath.s il.s error.s
+gcc -o ../9cc charutil.s fileutil.s list.s main.s parse.s parseutil.s stack.s tokenize.s typeutil.s varutil.s analyze.s mymath.s il.s abst.s gen_x8664.s gen_riscv.s error.s
